windows : fix test failed on race_disable.go
